[rbd] Clone images in iso format
In addition to "raw" images, also clone images in "iso" format. The
"iso" format is a special case of a raw image which can be cloned and
directly used as a volume.

Before the image inspection fixes for CVE-2024-32498 ISO images could be
uploaded as "raw" and clone worked as expected. Since these changes
uploads of images with an ISO 9660 header as "raw" images are rejected.

Change-Id: Id3b7e519849f672df578ffb5ff262e5683eec38c
Signed-off-by: Gaudenz Steinlin <gaudenz.steinlin@cloudscale.ch>